# src/main/java/com/sprint/findex/apidocs/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Findex API",
    "description": "가볍고 빠른 외부 API 연동 금융 분석 도구 API 문서",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://sprint-project-1196140422.ap-northeast-2.elb.amazonaws.com/sb/findex",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/index-data/{id}": {
      "delete": {
        "tags": [
          "지수 데이터 API"
        ],
        "summary": "지수 데이터 삭제",
        "description": "지수 데이터를 삭제합니다.",
        "operationId": "deleteIndexData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "지수 데이터 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "지수 데이터 삭제 성공"
          },
          "404": {
            "description": "삭제할 지수 데이터를 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "지수 데이터 API"
        ],
        "summary": "지수 데이터 수정",
        "description": "기존 지수 데이터를 수정합니다.",
        "operationId": "updateIndexData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "지수 데이터 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexDataUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "수정할 지수 데이터를 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "지수 데이터 수정 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IndexDataDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (유효하지 않은 데이터 값 등)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/index-data/{id}/chart": {
      "get": {
        "tags": [
          "지수 데이터 API"
        ],
        "summary": "지수 차트 조회",
        "description": "지수의 차트 데이터를 조회합니다.",
        "operationId": "getIndexChart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "지수 정보 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "periodType",
            "in": "query",
            "description": "차트 기간 유형 (DAILY, WEEKLY, MONTHLY)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DAILY",
              "description": "차트 기간 유형",
              "enum": [
                "MONTHLY",
                "QUARTERLY",
                "YEARLY"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "차트 데이터 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IndexChartDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (유효하지 않은 기간 유형 등)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "지수 정보를 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/index-data/performance/rank": {
      "get": {
        "tags": [
          "지수 데이터 API"
        ],
        "summary": "지수 성과 랭킹 조회",
        "description": "지수의 성과 분석 랭킹을 조회합니다.",
        "operationId": "getIndexPerformanceRank",
        "parameters": [
          {
            "name": "indexInfoId",
            "in": "query",
            "description": "지수 정보 ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "periodType",
            "in": "query",
            "description": "성과 기간 유형 (DAILY, WEEKLY, MONTHLY)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DAILY",
              "description": "단위 기간 유형",
              "enum": [
                "DAILY",
                "WEEKLY",
                "MONTHLY"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "최대 랭킹 수",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "400": {
            "description": "잘못된 요청 (유효하지 않은 기간 유형 등)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "성과 랭킹 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RankedIndexPerformanceDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/index-data/performance/favorite": {
      "get": {
        "tags": [
          "지수 데이터 API"
        ],
        "summary": "관심 지수 성과 조회",
        "description": "즐겨찾기로 등록된 지수들의 성과를 조회합니다.",
        "operationId": "getFavoriteIndexPerformance",
        "parameters": [
          {
            "name": "periodType",
            "in": "query",
            "description": "성과 기간 유형 (DAILY, WEEKLY, MONTHLY)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DAILY",
              "description": "단위 기간 유형",
              "enum": [
                "DAILY",
                "WEEKLY",
                "MONTHLY"
              ]
            }
          }
        ],
        "responses": {
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "관심 지수 성과 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IndexPerformanceDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/index-data/export/csv": {
      "get": {
        "tags": [
          "지수 데이터 API"
        ],
        "summary": "지수 데이터 CSV export",
        "description": "지수 데이터를 CSV 파일로 export합니다.",
        "operationId": "downloadIndexData",
        "parameters": [
          {
            "name": "indexInfoId",
            "in": "query",
            "description": "지수 정보 ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "시작 일자",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "종료 일자",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "sortField",
            "in": "query",
            "description": "정렬 필드 (baseDate, marketPrice, closingPrice, highPrice, lowPrice, versus, fluctuationRate, tradingQuantity, tradingPrice, marketTotalAmount)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "baseDate"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "정렬 방향 (asc, desc)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV 파일 생성 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (유효하지 않은 필터 값 등)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "에러 응답",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "에러 발생 시간",
            "example": "2025-03-06T05:39:06.152068Z"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP 상태 코드",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "에러 메시지",
            "example": "잘못된 요청입니다."
          },
          "details": {
            "type": "string",
            "description": "에러 상세 내용",
            "example": "부서 코드는 필수입니다."
          }
        }
      },
...
      "IndexDataDto": {
        "type": "object",
        "description": "지수 데이터 DTO",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "지수 데이터 ID",
            "example": 1
          },
          "indexInfoId": {
            "type": "integer",
            "format": "int64",
            "description": "지수 정보 ID",
            "example": 1
          },
          "baseDate": {
            "type": "string",
            "format": "date",
            "description": "기준 일자",
            "example": "2023-01-01"
          },
          "sourceType": {
            "type": "string",
            "description": "출처 (사용자, Open API)",
            "enum": [
              "USER",
              "OPEN_API"
            ],
            "example": "OPEN_API"
          },
          "marketPrice": {
            "type": "number",
            "description": "시가",
            "example": 2800.25
          },
          "closingPrice": {
            "type": "number",
            "description": "종가",
            "example": 2850.75
          },
          "highPrice": {
            "type": "number",
            "description": "고가",
            "example": 2870.5
          },
          "lowPrice": {
            "type": "number",
            "description": "저가",
            "example": 2795.3
          },
          "versus": {
            "type": "number",
            "description": "전일 대비 등락",
            "example": 50.5
          },
          "fluctuationRate": {
            "type": "number",
            "description": "전일 대비 등락률",
            "example": 1.8
          },
          "tradingQuantity": {
            "type": "integer",
            "format": "int64",
            "description": "거래량",
            "example": 1250000
          },
          "tradingPrice": {
            "type": "integer",
            "format": "int64",
            "description": "거래대금",
            "example": 3500000000
          },
          "marketTotalAmount": {
            "type": "integer",
            "format": "int64",
            "description": "상장 시가 총액",
            "example": 450000000000
          }
        }
      },
...
      "IndexDataUpdateRequest": {
        "type": "object",
        "description": "지수 데이터 수정 요청",
        "properties": {
          "marketPrice": {
            "type": "number",
            "description": "시가",
            "example": 2800.25
          },
          "closingPrice": {
            "type": "number",
            "description": "종가",
            "example": 2850.75
          },
          "highPrice": {
            "type": "number",
            "description": "고가",
            "example": 2870.5
          },
          "lowPrice": {
            "type": "number",
            "description": "저가",
            "example": 2795.3
          },
          "versus": {
            "type": "number",
            "description": "전일 대비 등락",
            "example": 50.5
          },
          "fluctuationRate": {
            "type": "number",
            "description": "전일 대비 등락률",
            "example": 1.8
          },
          "tradingQuantity": {
            "type": "integer",
            "format": "int64",
            "description": "거래량",
            "example": 1250000
          },
          "tradingPrice": {
            "type": "integer",
            "format": "int64",
            "description": "거래대금",
            "example": 3500000000
          },
          "marketTotalAmount": {
            "type": "integer",
            "format": "int64",
            "description": "상장 시가 총액",
            "example": 450000000000
          }
        }
      },
...
      "ChartDataPoint": {
        "type": "object",
        "description": "차트 데이터 포인트",
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "description": "날짜",
            "example": "2023-01-01"
          },
          "value": {
            "type": "number",
            "description": "값",
            "example": 2850.75
          }
        }
      },
      "IndexChartDto": {
        "type": "object",
        "description": "지수 차트 데이터 DTO",
        "properties": {
          "indexInfoId": {
            "type": "integer",
            "format": "int64",
            "description": "지수 정보 ID",
            "example": 1
          },
          "indexClassification": {
            "type": "string",
            "description": "지수 분류명",
            "example": "KOSPI시리즈"
          },
          "indexName": {
            "type": "string",
            "description": "지수명",
            "example": "IT 서비스"
          },
          "periodType": {
            "type": "string",
            "description": "차트 기간 유형 (DAILY, WEEKLY, MONTHLY)",
            "enum": [
              "MONTHLY",
              "QUARTERLY",
              "YEARLY"
            ],
            "example": "DAILY"
          },
          "dataPoints": {
            "type": "array",
            "description": "차트 데이터 목록",
            "items": {
              "$ref": "#/components/schemas/ChartDataPoint"
            }
          },
          "ma5DataPoints": {
            "type": "array",
            "description": "5일 이동평균선 데이터 목록",
            "items": {
              "$ref": "#/components/schemas/ChartDataPoint"
            }
          },
          "ma20DataPoints": {
            "type": "array",
            "description": "20일 이동평균선 데이터 목록",
            "items": {
              "$ref": "#/components/schemas/ChartDataPoint"
            }
          }
        }
      },
      "IndexPerformanceDto": {
        "type": "object",
        "description": "지수 성과 정보 DTO",
        "properties": {
          "indexInfoId": {
            "type": "integer",
            "format": "int64",
            "description": "지수 정보 ID",
            "example": 1
          },
          "indexClassification": {
            "type": "string",
            "description": "지수 분류명",
            "example": "KOSPI시리즈"
          },
          "indexName": {
            "type": "string",
            "description": "지수명",
            "example": "IT 서비스"
          },
          "versus": {
            "type": "number",
            "description": "단위 기간 대비 등락",
            "example": 50.5
          },
          "fluctuationRate": {
            "type": "number",
            "description": "단위 기간 대비 등락률",
            "example": 1.8
          },
          "currentPrice": {
            "type": "number",
            "description": "현재가",
            "example": 2850.75
          },
          "beforePrice": {
            "type": "number",
            "description": "단위 기간 전 값",
            "example": 2850.75
          }
        }
      },
      "RankedIndexPerformanceDto": {
        "type": "object",
        "description": "순위가 포함된 지수 성과 정보 DTO",
        "properties": {
          "performance": {
            "$ref": "#/components/schemas/IndexPerformanceDto",
            "description": "지수 성과 정보"
          },
          "rank": {
            "type": "integer",
            "format": "int32",
            "description": "순위",
            "example": 1
          }
        }
      },
...
    }
  }
}