# src/main/java/com/sprint/findex/dto/request/IndexDataUpdateRequest.java
package com.sprint.findex.dto.request;

import java.math.BigDecimal;

public record IndexDataUpdateRequest(
    BigDecimal marketPrice,
    BigDecimal closingPrice,
    BigDecimal highPrice,
    BigDecimal lowPrice,
    BigDecimal versus,
    BigDecimal fluctuationRate,
    Long tradingQuantity,
    Long tradingPrice,
    Long marketTotalAmount
) {

}